requirements.txt
# CAMS Voice AI POC — mic demo
numpy>=1.26.0
scipy>=1.11.0
soundfile>=0.12.0
PyAudio>=0.2.14
httpx>=0.27.0
websocket-client>=1.6.0
python-dotenv>=1.0.0
openai>=1.40.0
pydantic>=2.0.0
torch>=2.2.0
torchaudio>=2.2.0
sentence-transformers>=2.7.0
cartesia>=3.0.0
